Remove dist/ from repo
This adds dist/ to the gitignore, removing the need to build before submitting a PR and making merge conflicts easier to resolve.

This comes at the cost of bower support, since bower relied on these checked in resources. Existing tagged commits still exist with the bower.json file intact and the #npm branch can still be used as a latest bower.json point of access.